+++ src/DB/index.js
@@ -1,1 +1,2 @@
-export {default, valToSql, sql} from './DB'
+export {default} from './DB'
+export {default as SQLite, valToSql, sql} from './SQLite'