+++ src/index.js
@@ -1,4 +1,4 @@
-export {default as DB} from './DB'
+export {default as DB, SQLite} from './DB'
 export {default as EventQueue} from './EventQueue'
 export {default as EventSourcingDB} from './EventSourcingDB'
 export {default as JsonModel} from './JsonModel'